First batch of fixups for Linux/alpha Original rev  : rc@rc0.org.uk--2004-mairix/mairix--dev--0.15--patch-18
Two main areas of problems:

1. casting between pointers to 4-byte and 8-byte integers.

2. ctype functions unsafe when negative values passed as arguments (i.e. if
value comes straight from a (signed) char array or pointer dereference.)

Patches applied:

 * rc@rc0.org.uk--2004/mairix--alpha64--040629--base-0
   tag of rc@rc0.org.uk--2004-mairix/mairix--dev--0.15--patch-8

 * rc@rc0.org.uk--2004/mairix--alpha64--040629--patch-1
   Fix inconsistency in type of len arg to create_ro_mapping

 * rc@rc0.org.uk--2004/mairix--alpha64--040629--patch-2
   Fix up other callers of create_ro_mapping wrt 3rd arg type

 * rc@rc0.org.uk--2004/mairix--alpha64--040629--patch-3
   Make args to ctype.h functions safe wrt sign extension from char->int

+++ mbox.c
@@ -784,11 +784,11 @@ void build_mbox_lists(struct database *db, const char *folder_base, const char *
         mb->n_old_msgs_valid = mb->n_msgs;
       } else {
         unsigned char *va;
-        size_t len;
+        int len;
         create_ro_mapping(mb->path, &va, &len);
         if (va) {
           rescan_mbox(mb, va, len);
-          munmap(va, len);
+          munmap(va, (size_t) len);
         } else {
           /* Treat as dead mbox */
           deaden_mbox(mb);
@@ -807,7 +807,7 @@ int add_mbox_messages(struct database *db)/*{{{*/
   int i, j;
   int any_new = 0;
   unsigned char *va;
-  size_t valen;
+  int valen;
 
   for (i=0; i<db->n_mboxen; i++) {
     struct mbox *mb = &db->mboxen[i];
@@ -848,7 +848,7 @@ int add_mbox_messages(struct database *db)/*{{{*/
       ++db->n_msgs;
       any_new = 1;
     }
-    if (va) munmap(va, valen);
+    if (va) munmap(va, (size_t) valen);
   }
   return any_new;
 }
